refactor: type domain identities and provider semantics

+++ synctv-realtime/src/sync/realtime_manager/tests.rs
@@ -190,7 +190,7 @@ async fn test_local_critical_broadcast_does_not_panic_on_current_thread_runtime(
         username: "leaver".to_string(),
         remark_name: String::new(),
         display_tag: String::new(),
-        role: i32::from(synctv_core::models::RoomRole::Member),
+        role: synctv_core::models::RoomRole::Member,
         timestamp: synctv_core::SystemClock.now(),
     };
 
@@ -1253,7 +1253,7 @@ async fn test_publish_only_enqueues_redis_without_rebroadcasting_locally() -> Te
         username: "publish-only".to_string(),
         remark_name: String::new(),
         display_tag: String::new(),
-        role: i32::from(synctv_core::models::RoomRole::Member),
+        role: synctv_core::models::RoomRole::Member,
         timestamp: synctv_core::SystemClock.now(),
     };
